requirements.txt
python :3.6 +

Libraries :-

Django :
		Django==3.1
		djangorestframework==3.11.1
		
Mysql :
		mysql==0.0.2
		mysql-connector-python==8.0.23
		mysqlclient==2.0.3
		
Faker==4.1.1
		
Mysql Database Name : audio_project
